Initial Upload
Uploading the current, yet unusable state of the project. 

The WPF IDE is very different from Windows Forms and extremely different then any Game Engine and with the addition of the MVVM pattern, this project was way beyond my capabilities. Maybe after some time I'll be able to return and finish this project, after I learn WPF and MVVM.